ai-service/app/services/model_service.py
from PIL import Image
from typing import List
from loguru import logger

from app.core.config import settings


class ModelService:
    """AI model service for virtual try-on generation."""
    
    def __init__(self):
        self.model = None
        self.model_loaded = False
        self._load_model()
    
    def _load_model(self):
        """Load the try-on model."""
        try:
            # TODO: Load actual model (VITON-HD, HR-VITON, etc.)
            # self.model = load_model(settings.MODEL_PATH)
            logger.info("Model service initialized (placeholder mode)")
            self.model_loaded = True
        except Exception as e:
            logger.error(f"Failed to load model: {e}")
            self.model_loaded = False
    
    def generate_tryon(
        self, 
        user_image: Image.Image, 
        product_images: List[Image.Image]
    ) -> Image.Image:
        """
        Generate a virtual try-on image.
        
        This is a placeholder implementation. In production, this would:
        1. Extract pose/body information from user image
        2. Process clothing items from product images
        3. Use a neural network to generate the try-on result
        
        Popular models for this task:
        - VITON-HD
        - HR-VITON
        - DCI-VTON
        - Stable Diffusion + ControlNet
        """
        if not self.model_loaded:
            logger.warning("Model not loaded, returning placeholder result")
        
        # Placeholder: combine images
        # In real implementation, run through the try-on model
        
        # For now, return the user image with a simple overlay
        result = user_image.copy()
        
        # Add a subtle indicator that this is a try-on result
        # In production, this would be the generated try-on image
        
        logger.info("Generated try-on image (placeholder)")
        return result
    
    def is_ready(self) -> bool:
        """Check if the model is ready for inference."""
        return self.model_loaded
