ai-service/requirements.txt
# FastAPI AI Try-On Service
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3

# AWS S3 / Storage
boto3==1.34.14

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2

# AI/ML (optional - uncomment as needed)
# torch==2.1.2
# torchvision==0.16.2
# transformers==4.36.2
